Change logic for destination of default settings
Checking if the user installing kernelconfig has write access to /etc is
not such a good idea after all. This breaks installing in Gentoo with
standard mechanisms of the distutils-r1 eclass. In the end, if the user
does not have write access to /etc they should be making a --user with
pip.