Don't use stdout to send the base .config back
Previously the source scripts sent the the base config back to the main
program by passing it to stdout. Nice idea but soon the use of macros
will make that clunky. The location of the kernel sources directory is
now passed as an argument to the source script which will write the
.config file directly.